requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.6.0
django-filter==24.3
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
jiter==0.9.0
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
openai==1.68.2
packaging==24.1
pillow==11.0.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
sympy==1.13.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
